refactor : 최상위 view를 abstract class로 변경
