+++ src/main/java/ladder/view/input/InputView.java
@@ -1,0 +1,9 @@
+package ladder.view.input;
+
+import java.util.Scanner;
+
+public abstract class InputView {
+
+  protected static final Scanner SCANNER = new Scanner(System.in);
+
+}
